feat(core): implementar arquitectura multitenant y sistema de suscripciones
- Integración del paquete spatie/laravel-multitenancy.
- Creación de tablas core (tenants, plans, features, subscriptions).
- Seeder inicial para migrar datos existentes al Tenant 1 (Fase2Spa).
- Trait BelongsToTenant y Global Scope para aislamiento de datos automático.
- Middleware CheckSubscription para restringir el acceso a planes caducados.
- Pruebas unitarias para validar aislamiento de datos y middleware.

+++ database/migrations/2026_06_26_010640_create_turn_queues_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('turn_queues', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('schedule_id')->nullable();
+            $table->unsignedInteger('schedule_id')->nullable();
             $table->string('turn_number'); // e.g., 'A-1', 'B-1'
             $table->enum('status', ['waiting', 'in_progress', 'completed', 'delayed'])->default('waiting');
             $table->timestamp('notified_at')->nullable();
